+++ src/Map/AtlasMapRenderer.h
@@ -91,7 +91,7 @@ namespace OsmAnd
         virtual QVector<TileId> getVisibleTiles() const Q_DECL_OVERRIDE;
         virtual unsigned int getVisibleTilesCount() const Q_DECL_OVERRIDE;
         virtual PointD getCameraCoordinates() const Q_DECL_OVERRIDE;
-        virtual float getCameraHeight() const Q_DECL_OVERRIDE;
+        virtual double getCameraHeight() const Q_DECL_OVERRIDE;
 
         virtual int getTileSize3D() const Q_DECL_OVERRIDE;
